requirements.txt
openai
langchain
python-dotenv
streamlit-chat
tiktoken
pinecone-client